Updated: failed jobs file in advanced folder
